requirement.txt
# Football AI V2 Dependencies

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.1.0
supervision>=0.23.0
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.65.0
PyYAML>=6.0

# Roboflow
inference-gpu>=0.9.17  # Use inference-cpu if no GPU

# Sports analytics
sports @ git+https://github.com/roboflow/sports.git

# Team classification
transformers>=4.35.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Optional (for better performance)
accelerate>=0.24.0
